texto fixado, código orientado
